PHP Assignment 2
In this program, the printMultiplicationTable() function takes a number as a parameter and uses a for loop to iterate from 1 to 10. Inside the loop, it calculates the result of multiplying the given number with the loop counter ($number * $i) and then prints the multiplication expression and the result using the echo statement.

You can change the value of $givenNumber variable to the desired number for which you want to print the multiplication table. The program will output the multiplication table for that number.